requirements.txt
Django==1.5.1
South==0.7.6
anyjson==0.3.3
celery==3.0.22
django-celery==3.0.21
django-reversion==1.7.1
django-social-auth==0.7.23
gdata==2.0.17
google-api-python-client==1.1
httplib2==0.8
kombu==2.5.13
oauth2==1.5.211
oauth2client==1.1
pillow==2.0.0
pyopenssl==0.13
python-dateutil==2.1
python-gflags==2.0
python-ldap==2.4.13
pytz==2013b
